pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coursework-autopilot"
version = "0.1.0"
description = "Local MCP server and companion skill for coursework requirement extraction."
requires-python = ">=3.11"
dependencies = [
  "langdetect>=1.0.9",
  "mcp[cli]>=1.17.0",
  "pypdf>=6.0.0",
  "python-docx>=1.1.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
